refactor(docs): switch dx
