+++ apps/docs/content/components/switch/colors.ts
@@ -1,17 +1,4 @@
-const App = `import {Switch} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <div className="flex gap-4">
-      <Switch defaultSelected color="default">Default</Switch>
-      <Switch defaultSelected color="primary">Primary</Switch>
-      <Switch defaultSelected color="secondary">Secondary</Switch>
-      <Switch defaultSelected color="success">Success</Switch>
-      <Switch defaultSelected color="warning">Warning</Switch>
-      <Switch defaultSelected color="danger">Danger</Switch>
-    </div> 
-  );
-}`;
+import App from "./colors.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
