+++ apps/docs/content/components/switch/controlled.ts
@@ -1,17 +1,4 @@
-const App = `import {Switch} from "@nextui-org/react";
-
-export default function App() {
-  const [isSelected, setIsSelected] = React.useState(true);
-
-  return (
-    <div className="flex flex-col gap-2">
-      <Switch isSelected={isSelected} onValueChange={setIsSelected}>
-        Airplane mode
-      </Switch>  
-      <p className="text-small text-default-500">Selected: {isSelected ? "true" : "false"}</p>
-    </div>
-  )  
-}`;
+import App from "./controlled.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
