+++ apps/docs/content/components/switch/disabled.raw.jsx
@@ -1,0 +1,9 @@
+import {Switch} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <Switch defaultSelected isDisabled>
+      Automatic updates
+    </Switch>
+  );
+}
