+++ apps/docs/content/components/switch/label.raw.jsx
@@ -1,0 +1,5 @@
+import {Switch} from "@nextui-org/react";
+
+export default function App() {
+  return <Switch defaultSelected>Automatic updates</Switch>;
+}
