+++ apps/docs/content/components/switch/sizes.ts
@@ -1,14 +1,4 @@
-const App = `import {Switch} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <div className="flex gap-4">
-      <Switch defaultSelected size="sm">Small</Switch>
-      <Switch defaultSelected size="md">Medium</Switch>
-      <Switch defaultSelected size="lg">Large</Switch>
-    </div>  
-  );
-}`;
+import App from "./sizes.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
